Moved Query builder into own file
