+++ snuba/admin/static/components/query_display/types.tsx
@@ -24,6 +24,8 @@ type QueryResult = {
   timestamp: number;
   column_names: QueryResultColumnMetadata;
   rows: [QueryResultRow];
+  trace_output?: string;
+  error?: string;
 };
 
 export { ClickhouseNodeData, QueryRequest, QueryResult };
